got rid of a bug with 32-bit int overflow

+++ btwn_central.cxx
@@ -23,16 +23,16 @@ namespace CTF {
   }
 }
 
-void btwn_cnt_fast(Matrix<int> A, int b, Vector<double> & v, int nbatches){
+void btwn_cnt_fast(Matrix<int> A, int64_t b, Vector<double> & v, int nbatches){
   World dw = *A.wrld;
-  int n = A.nrow;
+  int64_t n = A.nrow;
 
   Semiring<mpath> p = get_mpath_semiring();
   Monoid<cpath> cp = get_cpath_monoid();
 
 
-  for (int ib=0; ib<n && (nbatches == 0 || ib/b<nbatches); ib+=b){
-    int k = std::min(b, n-ib);
+  for (int64_t ib=0; ib<n && (nbatches == 0 || ib/b<nbatches); ib+=b){
+    int64_t k = std::min(b, n-ib);
 
     //initialize shortest mpath vectors from the next k sources to the corresponding columns of the adjacency matrices and loops with weight 0
     ((Transform<int>)([=](int& w){ w = 0; }))(A["ii"]);
@@ -71,7 +71,7 @@ void btwn_cnt_fast(Matrix<int> A, int b, Vector<double> & v, int nbatches){
     //compute centrality scores by propagating them backwards from the furthest nodes (reverse Bellman Ford)
     int nbr = 0;
     double sbr = MPI_Wtime();
-    for (int i=0; i<n; i++, nbr++){
+    for (int64_t i=0; i<n; i++, nbr++){
       Matrix<cpath> C(cB);
       cB.set_zero();
       CTF::Timer tbr("Brandes");
